extract write validation fns
towards basic dynamics pt. 1 (#124)
